+++ assets/javascripts/kitten/components/notifications/icon-badge.test.js
@@ -1,41 +1,19 @@
 import React from 'react'
+import renderer from 'react-test-renderer'
 import { IconBadge } from 'kitten/components/notifications/icon-badge'
 
 describe('<IconBadge />', () => {
+  let component
+
   describe('by default', () => {
-    const iconBadge = shallow(<IconBadge />)
-
-    it('is a <span />', () => {
-      expect(iconBadge.type()).toBe('span')
+    beforeEach(() => {
+      component = renderer
+        .create(<IconBadge children="Custom children" valid big huge />)
+        .toJSON()
     })
 
-    it('has default classes', () => {
-      expect(iconBadge.hasClass('k-IconBadge')).toBe(true)
-      expect(iconBadge.find('.k-IconBadge__content').exists()).toBe(true)
-    })
-  })
-
-  describe('with custom class', () => {
-    const iconBadge = shallow(<IconBadge className="custom__class" />)
-
-    it('has a custom class', () => {
-      expect(iconBadge.hasClass('custom__class')).toBe(true)
-    })
-  })
-
-  describe('with other prop', () => {
-    const iconBadge = shallow(<IconBadge aria-hidden="true" />)
-
-    it('has an aria-hidden attribute', () => {
-      expect(iconBadge.props()['aria-hidden']).toBe('true')
-    })
-  })
-
-  describe('with children', () => {
-    const iconBadge = shallow(<IconBadge>A+</IconBadge>)
-
-    it('has text', () => {
-      expect(iconBadge.text()).toBe('A+')
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
     })
   })
 })