+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Remove `hover` state on `Select-control` class.
+- Fix: Use `StyleRoot` properly in `Comment` component in order to use
+  media-queries.
 
 ## [21.1.0] - 2018-07-19