+++ CHANGELOG.md
@@ -5,8 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Remove `hover` state on `Select-control` class.
+- Fix: Update size on `ButtonQuestionMarkIcon` component.
+- Fix: Update `locationIcon` svg.
 - Fix: Use `StyleRoot` properly in `Comment` component in order to use
-  media-queries.
 
 ## [21.1.0] - 2018-07-19